firmware: fix diversion and copy
Firmware files were not copied if the destination dir didn't exist. In
order to fix this issue, ensure the destination folder is created before
any processing, and pass `PathBuf`'s to `squash_file()` in order to have
a more homogenous processing.

This commit also moves the diversion code to a dedicated function,
making the code more readable.