firmware: make the dynamic partition name configurable
On most Android devices, the dynamic partition is unique and named
`super`. However, some devices have shipped with a "classic" partition
scheme and were converted to using dynamic partitions through an OS
upgrade; in such cases, the dynamic partition can be one of the
pre-existing partitions, such as `system_a` for example.

In order to account for those differences, we must make the dynamic
partition name configurable, which involves 2 changes:
* remove any mention to `make-dynpart-mappings@.service` from our
  `systemd` service file
* add a new, optional `dynpart` config option; when set, `droid-juicer`
  will directly call `systemctl` to start the service with the expected
  partition name, trying both `_a` and `_b` versions whenever relevant

Fixes #4