firmware: implement arbitrary folder copy
In order to communicate with the DSPs, we need a whole bunch of files
scattered across a few different folders. Rather than listing them all
individually, we should copy the whole folders to an arbitrary location
on the rootfs.

This change adds a new `folders` top-level config entry, very similar to
the `firmware` entry, except it doesn't have an `origin` attribute, and
its `files` attribute is renamed `folders`. This means the `folders`
entry is an array of objects with the following attributes:
* `partition`: the vendor partition name (without slot suffix)
* `destination`: the destination folder (can be anywhere on the rootfs,
  not limited to `/lib/firmware` like for `firmware` entries
* `folders`: array of folders to be copied, each item being an object
  with the following attributes:
  * `name`: the folder full path on the origin partition; the copied
    folder will be named after the last component of this path
  * `rename`: optional, rename the folder after the copy